requirements.txt
streamlit
face_recognition
opencv-contrib-python
numpy
pandas
Pillow
